refactor: move main page component in index
