--- src/components/main-page/index.ts
@@ -1,1 +1,0 @@
-export * from './MainPage';
